requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
pyjwt
bcrypt
python-multipart
python-dotenv
jinja2
aiogram
boto3
boto3-stubs
mypy-boto3-s3
twilio
# faststream[rabbit]
